flowers grows in 4 seasons: CRUD operations with ApiRespons Json Done

+++ src/main/java/com/devglan/springboothibernatelogin/service/impl/CategoryServiceImp.java
@@ -2,8 +2,12 @@ package com.devglan.springboothibernatelogin.service.impl;
 
 import com.devglan.springboothibernatelogin.Exception.CategoryNotFoundException;
 import com.devglan.springboothibernatelogin.dao.CategoryRepository;
+import com.devglan.springboothibernatelogin.dto.ApiResponse;
 import com.devglan.springboothibernatelogin.dto.CategoryDto;
+import com.devglan.springboothibernatelogin.dto.LoginDto;
+import com.devglan.springboothibernatelogin.dto.SignUpDto;
 import com.devglan.springboothibernatelogin.model.Category;
+import com.devglan.springboothibernatelogin.model.User;
 import com.devglan.springboothibernatelogin.service.CategoryService;
 
 import org.springframework.beans.BeanUtils;
@@ -13,7 +17,7 @@ import org.springframework.stereotype.Service;
 import java.util.Optional;
 
 @Service
-public class CategoryServiceImp {
+public class CategoryServiceImp implements CategoryService {
 
     @Autowired
     CategoryRepository categoryRepository;
@@ -22,9 +26,32 @@ public class CategoryServiceImp {
         return categoryRepository.save(user);
     }
 
-    public void deleteCategory(Integer id) {
-        Optional<Category> categoryOptional = categoryRepository.findById(id);
+    public Optional<Category> getCategory(int id) {
+        Optional<Category> category = categoryRepository.findById(id);
+        if(!category.isPresent()) {
+            throw new RuntimeException("Category Not Found");
+        }
+        return category;
+    }
 
+    public void deleteCategory(Integer id) {
+        Optional<Category> category = categoryRepository.findById(id);
+        if(!category.isPresent()) {
+            throw new RuntimeException("Category Not Found");
+        }
+        else
             categoryRepository.deleteById(id);
+    }
+
+    //Update Category
+    public Category updateCatedory(int id, Category category) {
+        Optional<Category> editCategory= categoryRepository.findById(id);
+        if (!editCategory.isPresent()) {
+            throw new CategoryNotFoundException(id);
+        } else {
+            categoryRepository.findById(id);
+            System.out.println(category.getLableCategory());
+        }
+        return categoryRepository.save(category);
     }
 }
